Require ActiveSupport 7
The broadcast logger API changed in ActiveSupport 7.1, replacing
`Logger#broadcast` with `BroadcastLogger`. In order for TC to upgrade
to Rails 7.1, we need to update this gem to use the new API.

Since this is a breaking change, we'll bump the version to 2.0.0.

See https://github.com/rails/rails/commit/1fbd812c47f7ba840390942e56d9b2ebbc260901